proposal: Move all top-level methods to groups and correct documentation (#252)
## 📝 Description

This change moves all top-level methods
(`LinodeClient.nodebalancers(...)`, etc.) to groups to ensure
consistency with the rest of the project. Additionally, alias methods
have been created to prevent introducing any breaking changes and to
ensure backwards compatibility.

Top-level list methods are now implemented as `__call__(...)` methods
for group classes, and create functions have been moved to `create(...)`
methods inside of each group.

These groups have all been moved to a separate `groups` package which
significantly reduces the size of the `linode_client` package.

This change also introduces a few fixes for documentation issues:
- The sidebar width has been slightly increased to prevent content on
the sidebar from overlapping with page content
- The page width has been slightly increased to improve readability
- Groups that were previously missing from the documentation have since
been added

## ✔️ How to Test

Testing refactor changes:
```bash
tox
```

Testing docs changes:
```bash
sphinx-autobuild docs docs/build
```
