+++ docs/linode_api4/linode_client.rst
@@ -52,6 +52,42 @@ example::
 See :any:`LinodeClient` for more information on the naming of these groups,
 although generally they are named the same as the first word of the group.
 
+AccountGroup
+^^^^^^^^^^^^
+
+Includes methods for managing your account.
+
+.. autoclass:: linode_api4.linode_client.AccountGroup
+   :members:
+   :special-members:
+
+DatabaseGroup
+^^^^^^^^^^^^^
+
+Includes methods for managing Linode Managed Databases.
+
+.. autoclass:: linode_api4.linode_client.DatabaseGroup
+   :members:
+   :special-members:
+
+DomainGroup
+^^^^^^^^^^^
+
+Includes methods for managing Linode Domains.
+
+.. autoclass:: linode_api4.linode_client.DomainGroup
+   :members:
+   :special-members:
+
+ImageGroup
+^^^^^^^^^^
+
+Includes methods for managing Linode Images.
+
+.. autoclass:: linode_api4.linode_client.ImageGroup
+   :members:
+   :special-members:
+
 LinodeGroup
 ^^^^^^^^^^^
 
@@ -60,30 +96,16 @@ accessing and working with associated features.
 
 .. autoclass:: linode_api4.linode_client.LinodeGroup
    :members:
+   :special-members:
 
-AccountGroup
-^^^^^^^^^^^^
+LKE Group
+^^^^^^^^^
 
-Includes methods for managing your account.
+Includes methods for interacting with Linode Kubernetes Engine.
 
-.. autoclass:: linode_api4.linode_client.AccountGroup
+.. autoclass:: linode_api4.linode_client.LKEGroup
    :members:
-
-ProfileGroup
-^^^^^^^^^^^^
-
-Includes methods for managing your user.
-
-.. autoclass:: linode_api4.linode_client.ProfileGroup
-   :members:
-
-NetworkingGroup
-^^^^^^^^^^^^^^^
-
-Includes methods for managing your networking systems.
-
-.. autoclass:: linode_api4.linode_client.NetworkingGroup
-   :members:
+   :special-members:
 
 LongviewGroup
 ^^^^^^^^^^^^^
@@ -92,6 +114,25 @@ Includes methods for interacting with our Longview service.
 
 .. autoclass:: linode_api4.linode_client.LongviewGroup
    :members:
+   :special-members:
+
+NetworkingGroup
+^^^^^^^^^^^^^^^
+
+Includes methods for managing your networking systems.
+
+.. autoclass:: linode_api4.linode_client.NetworkingGroup
+   :members:
+   :special-members:
+
+NodeBalancerGroup
+^^^^^^^^^^^^^^^^^
+
+Includes methods for managing Linode NodeBalancers.
+
+.. autoclass:: linode_api4.linode_client.NodeBalancerGroup
+   :members:
+   :special-members:
 
 ObjectStorageGroup
 ^^^^^^^^^^^^^^^^^^
@@ -101,16 +142,27 @@ with buckets and objects, use the s3 API directly with a library like `boto3`_.
 
 .. autoclass:: linode_api4.linode_client.ObjectStorageGroup
    :members:
+   :special-members:
 
 .. _boto3: https://github.com/boto/boto3
 
-LKE Group
-^^^^^^^^^
+ProfileGroup
+^^^^^^^^^^^^
 
-Includes methods for interacting with Linode Kubernetes Engine.
+Includes methods for managing your user.
 
-.. autoclass:: linode_api4.linode_client.LKEGroup
+.. autoclass:: linode_api4.linode_client.ProfileGroup
    :members:
+   :special-members:
+
+RegionGroup
+^^^^^^^^^^^
+
+Includes methods for accessing information about Linode Regions.
+
+.. autoclass:: linode_api4.linode_client.RegionGroup
+   :members:
+   :special-members:
 
 SupportGroup
 ^^^^^^^^^^^^
@@ -119,3 +171,22 @@ Includes methods for viewing and opening tickets with our support department.
 
 .. autoclass:: linode_api4.linode_client.SupportGroup
    :members:
+   :special-members:
+
+TagGroup
+^^^^^^^^
+
+Includes methods for managing Linode Tags.
+
+.. autoclass:: linode_api4.linode_client.TagGroup
+   :members:
+   :special-members:
+
+VolumeGroup
+^^^^^^^^^^^
+
+Includes methods for managing Linode Volumes.
+
+.. autoclass:: linode_api4.linode_client.VolumeGroup
+   :members:
+   :special-members:
