+++ linode_api4/groups/group.py
@@ -1,0 +1,3 @@
+class Group:
+    def __init__(self, client: "LinodeClient"):
+        self.client = client
