+++ linode_api4/linode_client.py
@@ -2,15 +2,14 @@ from __future__ import annotations
 
 import json
 import logging
-import os
 import time
-from datetime import datetime
 from typing import BinaryIO, Tuple
 
 import pkg_resources
 import requests
 
 from linode_api4.errors import ApiError, UnexpectedResponseError
+from linode_api4.groups import *
 from linode_api4.objects import *
 from linode_api4.objects.filtering import Filter
 
@@ -21,1940 +20,6 @@ from .util import drop_null_keys
 package_version = pkg_resources.require("linode_api4")[0].version
 
 logger = logging.getLogger(__name__)
-
-
-class Group:
-    def __init__(self, client: LinodeClient):
-        self.client = client
-
-
-class LinodeGroup(Group):
-    """
-    Encapsulates Linode-related methods of the :any:`LinodeClient`.  This
-    should not be instantiated on its own, but should instead be used through
-    an instance of :any:`LinodeClient`::
-
-       client = LinodeClient(token)
-       instances = client.linode.instances() # use the LinodeGroup
-
-    This group contains all features beneath the `/linode` group in the API v4.
-    """
-
-    def types(self, *filters):
-        """
-        Returns a list of Linode Instance types.  These may be used to create
-        or resize Linodes, or simply referenced on their own.  Types can be
-        filtered to return specific types, for example::
-
-           standard_types = client.linode.types(Type.class == "standard")
-
-        API documentation: https://www.linode.com/docs/api/linode-types/#types-list
-
-        :param filters: Any number of filters to apply to the query.
-
-        :returns: A list of types that match the query.
-        :rtype: PaginatedList of Type
-        """
-        return self.client._get_and_filter(Type, *filters)
-
-    def instances(self, *filters):
-        """
-        Returns a list of Linode Instances on your account.  You may filter
-        this query to return only Linodes that match specific criteria::
-
-           prod_linodes = client.linode.instances(Instance.group == "prod")
-
-        API Documentation: https://www.linode.com/docs/api/linode-instances/#linodes-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Instances that matched the query.
-        :rtype: PaginatedList of Instance
-        """
-        return self.client._get_and_filter(Instance, *filters)
-
-    def stackscripts(self, *filters, **kwargs):
-        """
-        Returns a list of :any:`StackScripts<StackScript>`, both public and
-        private.  You may filter this query to return only
-        :any:`StackScripts<StackScript>` that match certain criteria.  You may
-        also request only your own private :any:`StackScripts<StackScript>`::
-
-           my_stackscripts = client.linode.stackscripts(mine_only=True)
-
-        API Documentation: https://www.linode.com/docs/api/stackscripts/#stackscripts-list
-
-        :param filters: Any number of filters to apply to this query.
-        :param mine_only: If True, returns only private StackScripts
-        :type mine_only: bool
-
-        :returns: A list of StackScripts matching the query.
-        :rtype: PaginatedList of StackScript
-        """
-        # python2 can't handle *args and a single keyword argument, so this is a workaround
-        if "mine_only" in kwargs:
-            if kwargs["mine_only"]:
-                new_filter = Filter({"mine": True})
-                if filters:
-                    filters = list(filters)
-                    filters[0] = filters[0] & new_filter
-                else:
-                    filters = [new_filter]
-
-            del kwargs["mine_only"]
-
-        if kwargs:
-            raise TypeError(
-                "stackscripts() got unexpected keyword argument '{}'".format(
-                    kwargs.popitem()[0]
-                )
-            )
-
-        return self.client._get_and_filter(StackScript, *filters)
-
-    def kernels(self, *filters):
-        """
-        Returns a list of available :any:`Kernels<Kernel>`.  Kernels are used
-        when creating or updating :any:`LinodeConfigs,LinodeConfig>`.
-
-        API Documentation: https://www.linode.com/docs/api/linode-instances/#kernels-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of available kernels that match the query.
-        :rtype: PaginatedList of Kernel
-        """
-        return self.client._get_and_filter(Kernel, *filters)
-
-    # create things
-    def instance_create(
-        self, ltype, region, image=None, authorized_keys=None, **kwargs
-    ):
-        """
-        Creates a new Linode Instance. This function has several modes of operation:
-
-        **Create an Instance from an Image**
-
-        To create an Instance from an :any:`Image`, call `instance_create` with
-        a :any:`Type`, a :any:`Region`, and an :any:`Image`.  All three of
-        these fields may be provided as either the ID or the appropriate object.
-        In this mode, a root password will be generated and returned with the
-        new Instance object.  For example::
-
-           new_linode, password = client.linode.instance_create(
-               "g6-standard-2",
-               "us-east",
-               image="linode/debian9")
-
-           ltype = client.linode.types().first()
-           region = client.regions().first()
-           image = client.images().first()
-
-           another_linode, password = client.linode.instance_create(
-               ltype,
-               region,
-               image=image)
-
-        **Create an Instance from StackScript**
-
-        When creating an Instance from a :any:`StackScript`, an :any:`Image` that
-        the StackScript support must be provided..  You must also provide any
-        required StackScript data for the script's User Defined Fields..  For
-        example, if deploying `StackScript 10079`_ (which deploys a new Instance
-        with a user created from keys on `github`_::
-
-           stackscript = StackScript(client, 10079)
-
-           new_linode, password = client.linode.instance_create(
-              "g6-standard-2",
-              "us-east",
-              image="linode/debian9",
-              stackscript=stackscript,
-              stackscript_data={"gh_username": "example"})
-
-        In the above example, "gh_username" is the name of a User Defined Field
-        in the chosen StackScript.  For more information on StackScripts, see
-        the `StackScript guide`_.
-
-        .. _`StackScript 10079`: https://www.linode.com/stackscripts/view/10079
-        .. _`github`: https://github.com
-        .. _`StackScript guide`: https://www.linode.com/docs/platform/stackscripts/
-
-        **Create an Instance from a Backup**
-
-        To create a new Instance by restoring a :any:`Backup` to it, provide a
-        :any:`Type`, a :any:`Region`, and the :any:`Backup` to restore.  You
-        may provide either IDs or objects for all of these fields::
-
-           existing_linode = Instance(client, 123)
-           snapshot = existing_linode.available_backups.snapshot.current
-
-           new_linode = client.linode.instance_create(
-               "g6-standard-2",
-               "us-east",
-               backup=snapshot)
-
-        **Create an empty Instance**
-
-        If you want to create an empty Instance that you will configure manually,
-        simply call `instance_create` with a :any:`Type` and a :any:`Region`::
-
-           empty_linode = client.linode.instance_create("g6-standard-2", "us-east")
-
-        When created this way, the Instance will not be booted and cannot boot
-        successfully until disks and configs are created, or it is otherwise
-        configured.
-
-        API Documentation: https://www.linode.com/docs/api/linode-instances/#linode-create
-
-        :param ltype: The Instance Type we are creating
-        :type ltype: str or Type
-        :param region: The Region in which we are creating the Instance
-        :type region: str or Region
-        :param image: The Image to deploy to this Instance. If this is provided
-                      and no root_pass is given, a password will be generated
-                      and returned along with the new Instance.
-        :type image: str or Image
-        :param stackscript: The StackScript to deploy to the new Instance.  If
-                            provided, "image" is required and must be compatible
-                            with the chosen StackScript.
-        :type stackscript: int or StackScript
-        :param stackscript_data: Values for the User Defined Fields defined in
-                                 the chosen StackScript.  Does nothing if
-                                 StackScript is not provided.
-        :type stackscript_data: dict
-        :param backup: The Backup to restore to the new Instance.  May not be
-                       provided if "image" is given.
-        :type backup: int of Backup
-        :param authorized_keys: The ssh public keys to install in the linode's
-                                /root/.ssh/authorized_keys file.  Each entry may
-                                be a single key, or a path to a file containing
-                                the key.
-        :type authorized_keys: list or str
-        :param label: The display label for the new Instance
-        :type label: str
-        :param group: The display group for the new Instance
-        :type group: str
-        :param booted: Whether the new Instance should be booted.  This will
-                       default to True if the Instance is deployed from an Image
-                       or Backup.
-        :type booted: bool
-        :param tags: A list of tags to apply to the new instance.  If any of the
-                     tags included do not exist, they will be created as part of
-                     this operation.
-        :type tags: list[str]
-        :param private_ip: Whether the new Instance should have private networking
-                           enabled and assigned a private IPv4 address.
-        :type private_ip: bool
-
-        :returns: A new Instance object, or a tuple containing the new Instance and
-                  the generated password.
-        :rtype: Instance or tuple(Instance, str)
-        :raises ApiError: If contacting the API fails
-        :raises UnexpectedResponseError: If the API response is somehow malformed.
-                                         This usually indicates that you are using
-                                         an outdated library.
-        """
-        ret_pass = None
-        if image and not "root_pass" in kwargs:
-            ret_pass = Instance.generate_root_password()
-            kwargs["root_pass"] = ret_pass
-
-        authorized_keys = load_and_validate_keys(authorized_keys)
-
-        if "stackscript" in kwargs:
-            # translate stackscripts
-            kwargs["stackscript_id"] = (
-                kwargs["stackscript"].id
-                if issubclass(type(kwargs["stackscript"]), Base)
-                else kwargs["stackscript"]
-            )
-            del kwargs["stackscript"]
-
-        if "backup" in kwargs:
-            # translate backups
-            kwargs["backup_id"] = (
-                kwargs["backup"].id
-                if issubclass(type(kwargs["backup"]), Base)
-                else kwargs["backup"]
-            )
-            del kwargs["backup"]
-
-        params = {
-            "type": ltype.id if issubclass(type(ltype), Base) else ltype,
-            "region": region.id if issubclass(type(region), Base) else region,
-            "image": (image.id if issubclass(type(image), Base) else image)
-            if image
-            else None,
-            "authorized_keys": authorized_keys,
-        }
-        params.update(kwargs)
-
-        result = self.client.post("/linode/instances", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating linode!", json=result
-            )
-
-        l = Instance(self.client, result["id"], result)
-        if not ret_pass:
-            return l
-        return l, ret_pass
-
-    def stackscript_create(
-        self, label, script, images, desc=None, public=False, **kwargs
-    ):
-        """
-        Creates a new :any:`StackScript` on your account.
-
-        API Documentation: https://www.linode.com/docs/api/stackscripts/#stackscript-create
-
-        :param label: The label for this StackScript.
-        :type label: str
-        :param script: The script to run when an :any:`Instance` is deployed with
-                       this StackScript.  Must begin with a shebang (#!).
-        :type script: str
-        :param images: A list of :any:`Images<Image>` that this StackScript
-                       supports.  Instances will not be deployed from this
-                       StackScript unless deployed from one of these Images.
-        :type images: list of Image
-        :param desc: A description for this StackScript.
-        :type desc: str
-        :param public: Whether this StackScript is public.  Defaults to False.
-                       Once a StackScript is made public, it may not be set
-                       back to private.
-        :type public: bool
-
-        :returns: The new StackScript
-        :rtype: StackScript
-        """
-        image_list = None
-        if type(images) is list or type(images) is PaginatedList:
-            image_list = [
-                d.id if issubclass(type(d), Base) else d for d in images
-            ]
-        elif type(images) is Image:
-            image_list = [images.id]
-        elif type(images) is str:
-            image_list = [images]
-        else:
-            raise ValueError(
-                "images must be a list of Images or a single Image"
-            )
-
-        script_body = script
-        if not script.startswith("#!"):
-            # it doesn't look like a stackscript body, let's see if it's a file
-            if os.path.isfile(script):
-                with open(script) as f:
-                    script_body = f.read()
-            else:
-                raise ValueError(
-                    "script must be the script text or a path to a file"
-                )
-
-        params = {
-            "label": label,
-            "images": image_list,
-            "is_public": public,
-            "script": script_body,
-            "description": desc if desc else "",
-        }
-        params.update(kwargs)
-
-        result = self.client.post("/linode/stackscripts", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating StackScript!", json=result
-            )
-
-        s = StackScript(self.client, result["id"], result)
-        return s
-
-
-class ProfileGroup(Group):
-    """
-    Collections related to your user.
-    """
-
-    def __call__(self):
-        """
-        Retrieve the acting user's Profile, containing information about the
-        current user such as their email address, username, and uid.  This is
-        intended to be called off of a :any:`LinodeClient` object, like this::
-
-           profile = client.profile()
-
-        API Documentation: https://www.linode.com/docs/api/profile/#profile-view
-
-        :returns: The acting user's profile.
-        :rtype: Profile
-        """
-        result = self.client.get("/profile")
-
-        if not "username" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when getting profile!", json=result
-            )
-
-        p = Profile(self.client, result["username"], result)
-        return p
-
-    def tokens(self, *filters):
-        """
-        Returns the Person Access Tokens active for this user.
-
-        API Documentation: https://www.linode.com/docs/api/profile/#personal-access-tokens-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of tokens that matches the query.
-        :rtype: PaginatedList of PersonalAccessToken
-        """
-        return self.client._get_and_filter(PersonalAccessToken, *filters)
-
-    def token_create(self, label=None, expiry=None, scopes=None, **kwargs):
-        """
-        Creates and returns a new Personal Access Token.
-
-        API Documentation: https://www.linode.com/docs/api/profile/#personal-access-token-create
-
-        :param label: The label of the new Personal Access Token.
-        :type label: str
-        :param expiry: When the new Personal Accses Token will expire.
-        :type expiry: datetime or str
-        :param scopes: A space-separated list of OAuth scopes for this token.
-        :type scopes: str
-
-        :returns: The new Personal Access Token.
-        :rtype: PersonalAccessToken
-        """
-        if label:
-            kwargs["label"] = label
-        if expiry:
-            if isinstance(expiry, datetime):
-                expiry = datetime.strftime(expiry, "%Y-%m-%dT%H:%M:%S")
-            kwargs["expiry"] = expiry
-        if scopes:
-            kwargs["scopes"] = scopes
-
-        result = self.client.post("/profile/tokens", data=kwargs)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Personal Access Token!",
-                json=result,
-            )
-
-        token = PersonalAccessToken(self.client, result["id"], result)
-        return token
-
-    def apps(self, *filters):
-        """
-        Returns the Authorized Applications for this user
-
-        API Documentation: https://www.linode.com/docs/api/profile/#authorized-apps-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Authorized Applications for this user
-        :rtype: PaginatedList of AuthorizedApp
-        """
-        return self.client._get_and_filter(AuthorizedApp, *filters)
-
-    def ssh_keys(self, *filters):
-        """
-        Returns the SSH Public Keys uploaded to your profile.
-
-        API Documentation: https://www.linode.com/docs/api/profile/#ssh-keys-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of SSH Keys for this profile.
-        :rtype: PaginatedList of SSHKey
-        """
-        return self.client._get_and_filter(SSHKey, *filters)
-
-    def ssh_key_upload(self, key, label):
-        """
-        Uploads a new SSH Public Key to your profile  This key can be used in
-        later Linode deployments.
-
-        API Documentation: https://www.linode.com/docs/api/profile/#ssh-key-add
-
-        :param key: The ssh key, or a path to the ssh key.  If a path is provided,
-                    the file at the path must exist and be readable or an exception
-                    will be thrown.
-        :type key: str
-        :param label: The name to give this key.  This is purely aesthetic.
-        :type label: str
-
-        :returns: The newly uploaded SSH Key
-        :rtype: SSHKey
-        :raises ValueError: If the key provided does not appear to be valid, and
-                            does not appear to be a path to a valid key.
-        """
-        if not key.startswith(SSH_KEY_TYPES):
-            # this might be a file path - look for it
-            path = os.path.expanduser(key)
-            if os.path.isfile(path):
-                with open(path) as f:
-                    key = f.read().strip()
-            if not key.startswith(SSH_KEY_TYPES):
-                raise ValueError("Invalid SSH Public Key")
-
-        params = {
-            "ssh_key": key,
-            "label": label,
-        }
-
-        result = self.client.post("/profile/sshkeys", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when uploading SSH Key!", json=result
-            )
-
-        ssh_key = SSHKey(self.client, result["id"], result)
-        return ssh_key
-
-
-class LKEGroup(Group):
-    """
-    Encapsulates LKE-related methods of the :any:`LinodeClient`.  This
-    should not be instantiated on its own, but should instead be used through
-    an instance of :any:`LinodeClient`::
-
-       client = LinodeClient(token)
-       instances = client.lke.clusters() # use the LKEGroup
-
-    This group contains all features beneath the `/lke` group in the API v4.
-    """
-
-    def versions(self, *filters):
-        """
-        Returns a :any:`PaginatedList` of :any:`KubeVersion` objects that can be
-        used when creating an LKE Cluster.
-
-        API Documentation: https://www.linode.com/docs/api/linode-kubernetes-engine-lke/#kubernetes-versions-list
-
-        :param filters: Any number of filters to apply to the query.
-
-        :returns: A Paginated List of kube versions that match the query.
-        :rtype: PaginatedList of KubeVersion
-        """
-        return self.client._get_and_filter(KubeVersion, *filters)
-
-    def clusters(self, *filters):
-        """
-        Returns a :any:`PaginagtedList` of :any:`LKECluster` objects that belong
-        to this account.
-
-        https://www.linode.com/docs/api/linode-kubernetes-engine-lke/#kubernetes-clusters-list
-
-        :param filters: Any number of filters to apply to the query.
-
-        :returns: A Paginated List of LKE clusters that match the query.
-        :rtype: PaginatedList of LKECluster
-        """
-        return self.client._get_and_filter(LKECluster, *filters)
-
-    def cluster_create(self, region, label, node_pools, kube_version, **kwargs):
-        """
-        Creates an :any:`LKECluster` on this account in the given region, with
-        the given label, and with node pools as described.  For example::
-
-           client = LinodeClient(TOKEN)
-
-           # look up Region and Types to use.  In this example I'm just using
-           # the first ones returned.
-           target_region = client.regions().first()
-           node_type = client.linode.types()[0]
-           node_type_2 = client.linode.types()[1]
-           kube_version = client.lke.versions()[0]
-
-           new_cluster = client.lke.cluster_create(
-               target_region,
-               "example-cluster",
-               [client.lke.node_pool(node_type, 3), client.lke.node_pool(node_type_2, 3)],
-               kube_version
-            )
-
-        API Documentation: https://www.linode.com/docs/api/linode-kubernetes-engine-lke/#kubernetes-cluster-create
-
-        :param region: The Region to create this LKE Cluster in.
-        :type region: Region or str
-        :param label: The label for the new LKE Cluster.
-        :type label: str
-        :param node_pools: The Node Pools to create.
-        :type node_pools: one or a list of dicts containing keys "type" and "count".  See
-                          :any:`node_pool` for a convenient way to create correctly-
-                          formatted dicts.
-        :param kube_version: The version of Kubernetes to use
-        :type kube_version: KubeVersion or str
-        :param kwargs: Any other arguments to pass along to the API.  See the API
-                       docs for possible values.
-
-        :returns: The new LKE Cluster
-        :rtype: LKECluster
-        """
-        pools = []
-        if not isinstance(node_pools, list):
-            node_pools = [node_pools]
-
-        for c in node_pools:
-            if isinstance(c, dict):
-                new_pool = {
-                    "type": c["type"].id
-                    if "type" in c and issubclass(type(c["type"]), Base)
-                    else c.get("type"),
-                    "count": c.get("count"),
-                }
-
-                pools += [new_pool]
-
-        params = {
-            "label": label,
-            "region": region.id if issubclass(type(region), Base) else region,
-            "node_pools": pools,
-            "k8s_version": kube_version.id
-            if issubclass(type(kube_version), Base)
-            else kube_version,
-        }
-        params.update(kwargs)
-
-        result = self.client.post("/lke/clusters", data=params)
-
-        if "id" not in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating LKE cluster!", json=result
-            )
-
-        return LKECluster(self.client, result["id"], result)
-
-    def node_pool(self, node_type, node_count):
-        """
-        Returns a dict that is suitable for passing into the `node_pools` array
-        of :any:`cluster_create`.  This is a convenience method, and need not be
-        used to create Node Pools.  For proper usage, see the docs for :any:`cluster_create`.
-
-        :param node_type: The type of node to create in this node pool.
-        :type node_type: Type or str
-        :param node_count: The number of nodes to create in this node pool.
-        :type node_count: int
-
-        :returns: A dict describing the desired node pool.
-        :rtype: dict
-        """
-        return {
-            "type": node_type,
-            "count": node_count,
-        }
-
-
-class LongviewGroup(Group):
-    """
-    Collections related to Linode Longview.
-    """
-
-    def clients(self, *filters):
-        """
-        Requests and returns a paginated list of LongviewClients on your
-        account.
-
-        API Documentation: https://www.linode.com/docs/api/longview/#longview-clients-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Longview Clients matching the given filters.
-        :rtype: PaginatedList of LongviewClient
-        """
-        return self.client._get_and_filter(LongviewClient, *filters)
-
-    def client_create(self, label=None):
-        """
-        Creates a new LongviewClient, optionally with a given label.
-
-        API Documentation: https://www.linode.com/docs/api/longview/#longview-client-create
-
-        :param label: The label for the new client.  If None, a default label based
-            on the new client's ID will be used.
-
-        :returns: A new LongviewClient
-
-        :raises ApiError: If a non-200 status code is returned
-        :raises UnexpectedResponseError: If the returned data from the api does
-            not look as expected.
-        """
-        result = self.client.post("/longview/clients", data={"label": label})
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Longview Client!",
-                json=result,
-            )
-
-        c = LongviewClient(self.client, result["id"], result)
-        return c
-
-    def subscriptions(self, *filters):
-        """
-        Requests and returns a paginated list of LongviewSubscriptions available
-
-        API Documentation: https://www.linode.com/docs/api/longview/#longview-subscriptions-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Longview Subscriptions matching the given filters.
-        :rtype: PaginatedList of LongviewSubscription
-        """
-        return self.client._get_and_filter(LongviewSubscription, *filters)
-
-
-class AccountGroup(Group):
-    """
-    Collections related to your account.
-    """
-
-    def __call__(self):
-        """
-        Retrieves information about the acting user's account, such as billing
-        information.  This is intended to be called off of the :any:`LinodeClient`
-        class, like this::
-
-           account = client.account()
-
-        API Documentation: https://www.linode.com/docs/api/account/#account-view
-
-        :returns: Returns the acting user's account information.
-        :rtype: Account
-        """
-        result = self.client.get("/account")
-
-        if not "email" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when getting account!", json=result
-            )
-
-        return Account(self.client, result["email"], result)
-
-    def events(self, *filters):
-        """
-        Lists events on the current account matching the given filters.
-
-        API Documentation: https://www.linode.com/docs/api/account/#events-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of events on the current account matching the given filters.
-        :rtype: PaginatedList of Event
-        """
-
-        return self.client._get_and_filter(Event, *filters)
-
-    def events_mark_seen(self, event):
-        """
-        Marks event as the last event we have seen.  If event is an int, it is treated
-        as an event_id, otherwise it should be an event object whose id will be used.
-
-        API Documentation: https://www.linode.com/docs/api/account/#event-mark-as-seen
-
-        :param event: The Linode event to mark as seen.
-        :type event: Event or int
-        """
-        last_seen = event if isinstance(event, int) else event.id
-        self.client.post(
-            "{}/seen".format(Event.api_endpoint),
-            model=Event(self.client, last_seen),
-        )
-
-    def settings(self):
-        """
-        Returns the account settings data for this acocunt.  This is not  a
-        listing endpoint.
-
-        API Documentation: https://www.linode.com/docs/api/account/#account-settings-view
-
-        :returns: The account settings data for this account.
-        :rtype: AccountSettings
-        """
-        result = self.client.get("/account/settings")
-
-        if not "managed" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when getting account settings!",
-                json=result,
-            )
-
-        s = AccountSettings(self.client, result["managed"], result)
-        return s
-
-    def invoices(self):
-        """
-        Returns Invoices issued to this account.
-
-        API Documentation: https://www.linode.com/docs/api/account/#invoices-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: Invoices issued to this account.
-        :rtype: PaginatedList of Invoice
-        """
-        return self.client._get_and_filter(Invoice)
-
-    def payments(self):
-        """
-        Returns a list of Payments made on this account.
-
-        API Documentation: https://www.linode.com/docs/api/account/#payments-list
-
-        :returns: A list of payments made on this account.
-        :rtype: PaginatedList of Payment
-        """
-        return self.client._get_and_filter(Payment)
-
-    def oauth_clients(self, *filters):
-        """
-        Returns the OAuth Clients associated with this account.
-
-        API Documentation: https://www.linode.com/docs/api/account/#oauth-clients-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of OAuth Clients associated with this account.
-        :rtype: PaginatedList of OAuthClient
-        """
-        return self.client._get_and_filter(OAuthClient, *filters)
-
-    def oauth_client_create(self, name, redirect_uri, **kwargs):
-        """
-        Creates a new OAuth client.
-
-        API Documentation: https://www.linode.com/docs/api/account/#oauth-client-create
-
-        :param name: The name of this application.
-        :type name: str
-        :param redirect_uri: The location a successful log in from https://login.linode.com should be redirected to for this client.
-        :type redirect_uri: str
-
-        :returns: The created OAuth Client.
-        :rtype: OAuthClient
-        """
-        params = {
-            "label": name,
-            "redirect_uri": redirect_uri,
-        }
-        params.update(kwargs)
-
-        result = self.client.post("/account/oauth-clients", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating OAuth Client!", json=result
-            )
-
-        c = OAuthClient(self.client, result["id"], result)
-        return c
-
-    def users(self, *filters):
-        """
-        Returns a list of users on this account.
-
-        API Documentation: https://www.linode.com/docs/api/account/#users-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of users on this account.
-        :rtype: PaginatedList of User
-        """
-        return self.client._get_and_filter(User, *filters)
-
-    def logins(self):
-        """
-        Returns a collection of successful logins for all users on the account during the last 90 days.
-
-        API Documentation: https://www.linode.com/docs/api/account/#user-logins-list-all
-
-        :returns: A list of Logins on this account.
-        :rtype: PaginatedList of Login
-        """
-
-        return self.client._get_and_filter(Login)
-
-    def maintenance(self):
-        """
-        Returns a collection of Maintenance objects for any entity a user has permissions to view. Cancelled Maintenance objects are not returned.
-
-        API Documentation: https://www.linode.com/docs/api/account/#user-logins-list-all
-
-        :returns: A list of Maintenance objects on this account.
-        :rtype: List of Maintenance objects as MappedObjects
-        """
-
-        result = self.client.get(
-            "{}/maintenance".format(Account.api_endpoint), model=self
-        )
-
-        return [MappedObject(**r) for r in result["data"]]
-
-    def payment_methods(self):
-        """
-        Returns a  list of Payment Methods for this Account.
-
-        API Documentation: https://www.linode.com/docs/api/account/#payment-methods-list
-
-        :returns: A list of Payment Methods on this account.
-        :rtype: PaginatedList of PaymentMethod
-        """
-
-        return self.client._get_and_filter(PaymentMethod)
-
-    def add_payment_method(self, data, is_default, type):
-        """
-        Adds a Payment Method to your Account with the option to set it as the default method.
-
-        API Documentation: https://www.linode.com/docs/api/account/#payment-method-add
-
-        :param data: An object representing the credit card information you have on file with
-                     Linode to make Payments against your Account.
-        :type data: dict
-
-        Example usage::
-           data = {
-                "card_number": "4111111111111111",
-                "expiry_month": 11,
-                "expiry_year": 2020,
-                "cvv": "111"
-            }
-
-        :param is_default: Whether this Payment Method is the default method for
-                           automatically processing service charges.
-        :type is_default: bool
-
-        :param type: The type of Payment Method. Enum: ["credit_card]
-        :type type: str
-        """
-
-        if type != "credit_card":
-            raise ValueError("Unknown Payment Method type: {}".format(type))
-
-        if (
-            "card_number" not in data
-            or "expiry_month" not in data
-            or "expiry_year" not in data
-            or "cvv" not in data
-            or not data
-        ):
-            raise ValueError("Invalid credit card info provided")
-
-        params = {"data": data, "type": type, "is_default": is_default}
-
-        resp = self.client.post(
-            "{}/payment-methods".format(Account.api_endpoint),
-            model=self,
-            data=params,
-        )
-
-        if "error" in resp:
-            raise UnexpectedResponseError(
-                "Unexpected response when adding payment method!",
-                json=resp,
-            )
-
-    def notifications(self):
-        """
-        Returns a collection of Notification objects representing important, often time-sensitive items related to your Account.
-
-        API Documentation: https://www.linode.com/docs/api/account/#notifications-list
-
-        :returns: A list of Notifications on this account.
-        :rtype: List of Notification objects as MappedObjects
-        """
-
-        result = self.client.get(
-            "{}/notifications".format(Account.api_endpoint), model=self
-        )
-
-        return [MappedObject(**r) for r in result["data"]]
-
-    def linode_managed_enable(self):
-        """
-        Enables Linode Managed for the entire account and sends a welcome email to the account’s associated email address.
-
-        API Documentation: https://www.linode.com/docs/api/account/#linode-managed-enable
-        """
-
-        resp = self.client.post(
-            "{}/settings/managed-enable".format(Account.api_endpoint),
-            model=self,
-        )
-
-        if "error" in resp:
-            raise UnexpectedResponseError(
-                "Unexpected response when enabling Linode Managed!",
-                json=resp,
-            )
-
-    def add_promo_code(self, promo_code):
-        """
-        Adds an expiring Promo Credit to your account.
-
-        API Documentation: https://www.linode.com/docs/api/account/#promo-credit-add
-
-        :param promo_code: The Promo Code.
-        :type promo_code: str
-        """
-
-        params = {
-            "promo_code": promo_code,
-        }
-
-        resp = self.client.post(
-            "{}/promo-codes".format(Account.api_endpoint),
-            model=self,
-            data=params,
-        )
-
-        if "error" in resp:
-            raise UnexpectedResponseError(
-                "Unexpected response when adding Promo Code!",
-                json=resp,
-            )
-
-    def service_transfers(self):
-        """
-        Returns a collection of all created and accepted Service Transfers for this account, regardless of the user that created or accepted the transfer.
-
-        API Documentation: https://www.linode.com/docs/api/account/#service-transfers-list
-
-        :returns: A list of Service Transfers on this account.
-        :rtype: PaginatedList of ServiceTransfer
-        """
-
-        return self.client._get_and_filter(ServiceTransfer)
-
-    def service_transfer_create(self, entities):
-        """
-        Creates a transfer request for the specified services.
-
-        API Documentation: https://www.linode.com/docs/api/account/#service-transfer-create
-
-        :param entities: A collection of the services to include in this transfer request, separated by type.
-        :type entities: dict
-
-        Example usage::
-           entities = {
-                "linodes": [
-                    111,
-                    222
-                ]
-            }
-        """
-
-        if not entities:
-            raise ValueError("Entities must be provided!")
-
-        bad_entries = [
-            k for k, v in entities.items() if not isinstance(v, list)
-        ]
-        if len(bad_entries) > 0:
-            raise ValueError(
-                f"Got unexpected type for entity lists: {', '.join(bad_entries)}"
-            )
-
-        params = {"entities": entities}
-
-        resp = self.client.post(
-            "{}/service-transfers".format(Account.api_endpoint),
-            model=self,
-            data=params,
-        )
-
-        if "error" in resp:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Service Transfer!",
-                json=resp,
-            )
-
-    def transfer(self):
-        """
-        Returns a MappedObject containing the account's transfer pool data.
-
-        API Documentation: https://www.linode.com/docs/api/account/#network-utilization-view
-
-        :returns: Information about this account's transfer pool data.
-        :rtype: MappedObject
-        """
-        result = self.client.get("/account/transfer")
-
-        if not "used" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when getting Transfer Pool!"
-            )
-
-        return MappedObject(**result)
-
-    def user_create(self, email, username, restricted=True):
-        """
-        Creates a new user on your account.  If you create an unrestricted user,
-        they will immediately be able to access everything on your account.  If
-        you create a restricted user, you must grant them access to parts of your
-        account that you want to allow them to manage (see :any:`User.grants` for
-        details).
-
-        The new user will receive an email inviting them to set up their password.
-        This must be completed before they can log in.
-
-        API Documentation: https://www.linode.com/docs/api/account/#user-create
-
-        :param email: The new user's email address.  This is used to finish setting
-                      up their user account.
-        :type email: str
-        :param username: The new user's unique username.  They will use this username
-                         to log in.
-        :type username: str
-        :param restricted: If True, the new user must be granted access to parts of
-                           the account before they can do anything.  If False, the
-                           new user will immediately be able to manage the entire
-                           account.  Defaults to True.
-        :type restricted: True
-
-        :returns The new User.
-        :rtype: User
-        """
-        params = {
-            "email": email,
-            "username": username,
-            "restricted": restricted,
-        }
-        result = self.client.post("/account/users", data=params)
-
-        if not all(
-            [c in result for c in ("email", "restricted", "username")]
-        ):  # pylint: disable=use-a-generator
-            raise UnexpectedResponseError(
-                "Unexpected response when creating user!", json=result
-            )
-
-        u = User(self.client, result["username"], result)
-        return u
-
-
-class NetworkingGroup(Group):
-    """
-    Collections related to Linode Networking.
-    """
-
-    def firewalls(self, *filters):
-        """
-        Retrieves the Firewalls your user has access to.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#firewalls-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Firewalls the acting user can access.
-        :rtype: PaginatedList of Firewall
-        """
-        return self.client._get_and_filter(Firewall, *filters)
-
-    def firewall_create(self, label, rules, **kwargs):
-        """
-        Creates a new Firewall, either in the given Region or
-        attached to the given Instance.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#firewall-create
-
-        :param label: The label for the new Firewall.
-        :type label: str
-        :param rules: The rules to apply to the new Firewall. For more information on Firewall rules, see our `Firewalls Documentation`_.
-        :type rules: dict
-
-        :returns: The new Firewall.
-        :rtype: Firewall
-
-        Example usage::
-
-           rules = {
-                'outbound': [
-                    {
-                        'action': 'ACCEPT',
-                        'addresses': {
-                            'ipv4': [
-                                '0.0.0.0/0'
-                            ],
-                            'ipv6': [
-                                "ff00::/8"
-                            ]
-                        },
-                        'description': 'Allow HTTP out.',
-                        'label': 'allow-http-out',
-                        'ports': '80',
-                        'protocol': 'TCP'
-                    }
-                ],
-                'outbound_policy': 'DROP',
-                'inbound': [],
-                'inbound_policy': 'DROP'
-            }
-
-            firewall = client.networking.firewall_create('my-firewall', rules)
-
-        .. _Firewalls Documentation: https://www.linode.com/docs/api/networking/#firewall-create__request-body-schema
-        """
-
-        params = {
-            "label": label,
-            "rules": rules,
-        }
-        params.update(kwargs)
-
-        result = self.client.post("/networking/firewalls", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Firewall!", json=result
-            )
-
-        f = Firewall(self.client, result["id"], result)
-        return f
-
-    def ips(self, *filters):
-        """
-        Returns a list of IP addresses on this account, excluding private addresses.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#ip-addresses-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of IP addresses on this account.
-        :rtype: PaginatedList of IPAddress
-        """
-        return self.client._get_and_filter(IPAddress, *filters)
-
-    def ipv6_ranges(self, *filters):
-        """
-        Returns a list of IPv6 ranges on this account.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#ipv6-ranges-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of IPv6 ranges on this account.
-        :rtype: PaginatedList of IPv6Range
-        """
-        return self.client._get_and_filter(IPv6Range, *filters)
-
-    def ipv6_pools(self, *filters):
-        """
-        Returns a list of IPv6 pools on this account.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#ipv6-pools-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of IPv6 pools on this account.
-        :rtype: PaginatedList of IPv6Pool
-        """
-
-        return self.client._get_and_filter(IPv6Pool, *filters)
-
-    def vlans(self, *filters):
-        """
-        .. note:: This endpoint is in beta. This will only function if base_url is set to `https://api.linode.com/v4beta`.
-
-        Returns a list of VLANs on your account.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#vlans-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A List of VLANs on your account.
-        :rtype: PaginatedList of VLAN
-        """
-        return self.client._get_and_filter(VLAN, *filters)
-
-    def ips_assign(self, region, *assignments):
-        """
-        Redistributes :any:`IP Addressees<IPAddress>` within a single region.
-        This function takes a :any:`Region` and a list of assignments to make,
-        then requests that the assignments take place.  If any :any:`Instance`
-        ends up without a public IP, or with more than one private IP, all of
-        the assignments will fail.
-
-        .. note::
-           This function *does not* update the local Linode Instance objects
-           when called.  In order to see the new addresses on the local
-           instance objects, be sure to invalidate them with ``invalidate()``
-           after this completes.
-
-        Example usage::
-
-           linode1 = Instance(client, 123)
-           linode2 = Instance(client, 456)
-
-           # swap IPs between linodes 1 and 2
-           client.networking.assign_ips(linode1.region,
-                                        linode1.ips.ipv4.public[0].to(linode2),
-                                        linode2.ips.ipv4.public[0].to(linode1))
-
-           # make sure linode1 and linode2 have updated ipv4 and ips values
-           linode1.invalidate()
-           linode2.invalidate()
-
-        API Documentation: https://www.linode.com/docs/api/networking/#linodes-assign-ipv4s
-
-        :param region: The Region in which the assignments should take place.
-                       All Instances and IPAddresses involved in the assignment
-                       must be within this region.
-        :type region: str or Region
-        :param assignments: Any number of assignments to make.  See
-                            :any:`IPAddress.to` for details on how to construct
-                            assignments.
-        :type assignments: dct
-
-        DEPRECATED: Use ip_addresses_assign() instead
-        """
-        for a in assignments:
-            if not "address" in a or not "linode_id" in a:
-                raise ValueError("Invalid assignment: {}".format(a))
-        if isinstance(region, Region):
-            region = region.id
-
-        self.client.post(
-            "/networking/ipv4/assign",
-            data={
-                "region": region,
-                "assignments": assignments,
-            },
-        )
-
-    def ip_allocate(self, linode, public=True):
-        """
-        Allocates an IP to a Instance you own.  Additional IPs must be requested
-        by opening a support ticket first.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#ip-address-allocate
-
-        :param linode: The Instance to allocate the new IP for.
-        :type linode: Instance or int
-        :param public: If True, allocate a public IP address.  Defaults to True.
-        :type public: bool
-
-        :returns: The new IPAddress.
-        :rtype: IPAddress
-        """
-        result = self.client.post(
-            "/networking/ips/",
-            data={
-                "linode_id": linode.id if isinstance(linode, Base) else linode,
-                "type": "ipv4",
-                "public": public,
-            },
-        )
-
-        if not "address" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when adding IPv4 address!", json=result
-            )
-
-        ip = IPAddress(self.client, result["address"], result)
-        return ip
-
-    def ips_share(self, linode, *ips):
-        """
-        Shares the given list of :any:`IPAddresses<IPAddress>` with the provided
-        :any:`Instance`.  This will enable the provided Instance to bring up the
-        shared IP Addresses even though it does not own them.
-
-        API Documentation: https://www.linode.com/docs/api/networking/#ipv4-sharing-configure
-
-        :param linode: The Instance to share the IPAddresses with.  This Instance
-                       will be able to bring up the given addresses.
-        :type: linode: int or Instance
-        :param ips: Any number of IPAddresses to share to the Instance.
-        :type ips: str or IPAddress
-
-        DEPRECATED: Use ip_addresses_share() instead
-        """
-        if not isinstance(linode, Instance):
-            # make this an object
-            linode = Instance(self.client, linode)
-
-        params = []
-        for ip in ips:
-            if isinstance(ip, str):
-                params.append(ip)
-            elif isinstance(ip, IPAddress):
-                params.append(ip.address)
-            else:
-                params.append(str(ip))  # and hope that works
-
-        params = {"ips": params}
-
-        self.client.post(
-            "{}/networking/ipv4/share".format(Instance.api_endpoint),
-            model=linode,
-            data=params,
-        )
-
-        linode.invalidate()  # clear the Instance's shared IPs
-
-    def ip_addresses_share(self, ips, linode):
-        """
-        Configure shared IPs. P sharing allows IP address reassignment
-        (also referred to as IP failover) from one Linode to another if the
-        primary Linode becomes unresponsive. This means that requests to the primary Linode’s
-        IP address can be automatically rerouted to secondary Linodes at the configured shared IP addresses.
-
-        :param linode: The id of the Instance or the Instance to share the IPAddresses with.
-                          This Instance will be able to bring up the given addresses.
-        :type: linode: int or Instance
-        :param ips: Any number of IPAddresses to share to the Instance.
-        :type ips: str or IPAddress
-        """
-
-        params = {
-            "ips": ips
-            if not isinstance(ips[0], IPAddress)
-            else [ip.address for ip in ips],
-            "linode_id": linode
-            if not isinstance(linode, Instance)
-            else linode.id,
-        }
-
-        self.client.post("/networking/ips/share", model=self, data=params)
-
-    def ip_addresses_assign(self, assignments, region):
-        """
-        Assign multiple IPv4 addresses and/or IPv6 ranges to multiple Linodes in one Region.
-        This allows swapping, shuffling, or otherwise reorganizing IPs to your Linodes.
-
-        The following restrictions apply:
-            - All Linodes involved must have at least one public IPv4 address after assignment.
-            - Linodes may have no more than one assigned private IPv4 address.
-            - Linodes may have no more than one assigned IPv6 range.
-
-
-        :param region: The Region in which the assignments should take place.
-                       All Instances and IPAddresses involved in the assignment
-                       must be within this region.
-        :type region: str or Region
-        :param assignments: Any number of assignments to make.  See
-                            :any:`IPAddress.to` for details on how to construct
-                            assignments.
-        :type assignments: dct
-        """
-
-        for a in assignments["assignments"]:
-            if not "address" in a or not "linode_id" in a:
-                raise ValueError("Invalid assignment: {}".format(a))
-
-        if isinstance(region, Region):
-            region = region.id
-
-        params = {"assignments": assignments, "region": region}
-
-        self.client.post("/networking/ips/assign", model=self, data=params)
-
-
-class SupportGroup(Group):
-    """
-    Collections related to support tickets.
-    """
-
-    def tickets(self, *filters):
-        """
-        Returns a list of support tickets on this account.
-
-        API Documentation: https://www.linode.com/docs/api/support/#support-tickets-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of support tickets on this account.
-        :rtype: PaginatedList of SupportTicket
-        """
-
-        return self.client._get_and_filter(SupportTicket, *filters)
-
-    def ticket_open(
-        self,
-        summary,
-        description,
-        managed_issue=False,
-        regarding=None,
-        **kwargs,
-    ):
-        """
-        Opens a support ticket on this account.
-
-        API Documentation: https://www.linode.com/docs/api/support/#support-ticket-open
-
-        :param summary: The summary or title for this support ticket.
-        :type summary: str
-        :param description: The full details of the issue or question.
-        :type description: str
-        :param regarding: The resource being referred to in this ticket.
-        :type regarding:
-        :param managed_issue: Designates if this ticket relates to a managed service.
-        :type managed_issue: bool
-
-        :returns: The new support ticket.
-        :rtype: SupportTicket
-        """
-        params = {
-            "summary": summary,
-            "description": description,
-            "managed_issue": managed_issue,
-        }
-
-        type_to_id = {
-            Instance: "linode_id",
-            Domain: "domain_id",
-            NodeBalancer: "nodebalancer_id",
-            Volume: "volume_id",
-            Firewall: "firewall_id",
-            LKECluster: "lkecluster_id",
-            Database: "database_id",
-            LongviewClient: "longviewclient_id",
-        }
-
-        params.update(kwargs)
-
-        if regarding:
-            id_attr = type_to_id.get(type(regarding))
-
-            if id_attr is not None:
-                params[id_attr] = regarding.id
-            elif isinstance(regarding, VLAN):
-                params["vlan"] = regarding.label
-                params["region"] = regarding.region
-            else:
-                raise ValueError(
-                    "Cannot open ticket regarding type {}!".format(
-                        type(regarding)
-                    )
-                )
-
-        result = self.client.post("/support/tickets", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating ticket!", json=result
-            )
-
-        t = SupportTicket(self.client, result["id"], result)
-        return t
-
-
-class ObjectStorageGroup(Group):
-    """
-    This group encapsulates all endpoints under /object-storage, including viewing
-    available clusters and managing keys.
-    """
-
-    def clusters(self, *filters):
-        """
-        Returns a list of available Object Storage Clusters.  You may filter
-        this query to return only Clusters that are available in a specific region::
-
-           us_east_clusters = client.object_storage.clusters(ObjectStorageCluster.region == "us-east")
-
-        API Documentation: https://www.linode.com/docs/api/object-storage/#clusters-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Object Storage Clusters that matched the query.
-        :rtype: PaginatedList of ObjectStorageCluster
-        """
-        return self.client._get_and_filter(ObjectStorageCluster, *filters)
-
-    def keys(self, *filters):
-        """
-        Returns a list of Object Storage Keys active on this account.  These keys
-        allow third-party applications to interact directly with Linode Object Storage.
-
-        API Documentation: https://www.linode.com/docs/api/object-storage/#object-storage-keys-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Object Storage Keys that matched the query.
-        :rtype: PaginatedList of ObjectStorageKeys
-        """
-        return self.client._get_and_filter(ObjectStorageKeys, *filters)
-
-    def keys_create(self, label, bucket_access=None):
-        """
-        Creates a new Object Storage keypair that may be used to interact directly
-        with Linode Object Storage in third-party applications.  This response is
-        the only time that "secret_key" will be populated - be sure to capture its
-        value or it will be lost forever.
-
-        If given, `bucket_access` will cause the new keys to be restricted to only
-        the specified level of access for the specified buckets.  For example, to
-        create a keypair that can only access the "example" bucket in all clusters
-        (and assuming you own that bucket in every cluster), you might do this::
-
-           client = LinodeClient(TOKEN)
-
-           # look up clusters
-           all_clusters = client.object_storage.clusters()
-
-           new_keys = client.object_storage.keys_create(
-               "restricted-keys",
-               bucket_access=[
-                   client.object_storage.bucket_access(cluster, "example", "read_write")
-                   for cluster in all_clusters
-               ],
-           )
-
-        To create a keypair that can only read from the bucket "example2" in the
-        "us-east-1" cluster (an assuming you own that bucket in that cluster),
-        you might do this::
-
-           client = LinodeClient(TOKEN)
-           new_keys_2 = client.object_storage.keys_create(
-               "restricted-keys-2",
-               bucket_access=client.object_storage.bucket_access("us-east-1", "example2", "read_only"),
-           )
-
-        API Documentation: https://www.linode.com/docs/api/object-storage/#object-storage-key-create
-
-        :param label: The label for this keypair, for identification only.
-        :type label: str
-        :param bucket_access: One or a list of dicts with keys "cluster,"
-                              "permissions", and "bucket_name".  If given, the
-                              resulting Object Storage keys will only have the
-                              requested level of access to the requested buckets,
-                              if they exist and are owned by you.  See the provided
-                              :any:`bucket_access` function for a convenient way
-                              to create these dicts.
-        :type bucket_access: dict or list of dict
-
-        :returns: The new keypair, with the secret key populated.
-        :rtype: ObjectStorageKeys
-        """
-        params = {"label": label}
-
-        if bucket_access is not None:
-            if not isinstance(bucket_access, list):
-                bucket_access = [bucket_access]
-
-            ba = [
-                {
-                    "permissions": c.get("permissions"),
-                    "bucket_name": c.get("bucket_name"),
-                    "cluster": c.id
-                    if "cluster" in c and issubclass(type(c["cluster"]), Base)
-                    else c.get("cluster"),
-                }
-                for c in bucket_access
-            ]
-
-            params["bucket_access"] = ba
-
-        result = self.client.post("/object-storage/keys", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Object Storage Keys!",
-                json=result,
-            )
-
-        ret = ObjectStorageKeys(self.client, result["id"], result)
-        return ret
-
-    def bucket_access(self, cluster, bucket_name, permissions):
-        """
-        Returns a dict formatted to be included in the `bucket_access` argument
-        of :any:`keys_create`.  See the docs for that method for an example of
-        usage.
-
-        :param cluster: The Object Storage cluster to grant access in.
-        :type cluster: :any:`ObjectStorageCluster` or str
-        :param bucket_name: The name of the bucket to grant access to.
-        :type bucket_name: str
-        :param permissions: The permissions to grant.  Should be one of "read_only"
-                            or "read_write".
-        :type permissions: str
-
-        :returns: A dict formatted correctly for specifying  bucket access for
-                  new keys.
-        :rtype: dict
-        """
-        return {
-            "cluster": cluster,
-            "bucket_name": bucket_name,
-            "permissions": permissions,
-        }
-
-    def cancel(self):
-        """
-        Cancels Object Storage service.  This may be a destructive operation.  Once
-        cancelled, you will no longer receive the transfer for or be billed for
-        Object Storage, and all keys will be invalidated.
-
-        API Documentation: https://www.linode.com/docs/api/object-storage/#object-storage-cancel
-        """
-        self.client.post("/object-storage/cancel", data={})
-        return True
-
-
-class DatabaseGroup(Group):
-    """
-    Encapsulates Linode Managed Databases related methods of the :any:`LinodeClient`. This
-    should not be instantiated on its own, but should instead be used through
-    an instance of :any:`LinodeClient`::
-
-       client = LinodeClient(token)
-       instances = client.database.instances() # use the DatabaseGroup
-
-    This group contains all features beneath the `/databases` group in the API v4.
-    """
-
-    def types(self, *filters):
-        """
-        Returns a list of Linode Database-compatible Instance types.
-        These may be used to create Managed Databases, or simply
-        referenced to on their own. DatabaseTypes can be
-        filtered to return specific types, for example::
-
-           database_types = client.database.types(DatabaseType.deprecated == False)
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-database-types-list
-
-        :param filters: Any number of filters to apply to the query.
-
-        :returns: A list of types that match the query.
-        :rtype: PaginatedList of DatabaseType
-        """
-        return self.client._get_and_filter(DatabaseType, *filters)
-
-    def engines(self, *filters):
-        """
-        Returns a list of Linode Managed Database Engines.
-        These may be used to create Managed Databases, or simply
-        referenced to on their own. Engines can be filtered to
-        return specific engines, for example::
-
-           mysql_engines = client.database.engines(DatabaseEngine.engine == 'mysql')
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-database-engines-list
-
-        :param filters: Any number of filters to apply to the query.
-
-        :returns: A list of types that match the query.
-        :rtype: PaginatedList of DatabaseEngine
-        """
-        return self.client._get_and_filter(DatabaseEngine, *filters)
-
-    def instances(self, *filters):
-        """
-        Returns a list of Managed Databases active on this account.
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-databases-list-all
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of databases that matched the query.
-        :rtype: PaginatedList of Database
-        """
-        return self.client._get_and_filter(Database, *filters)
-
-    def mysql_instances(self, *filters):
-        """
-        Returns a list of Managed MySQL Databases active on this account.
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-mysql-databases-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of MySQL databases that matched the query.
-        :rtype: PaginatedList of MySQLDatabase
-        """
-        return self.client._get_and_filter(MySQLDatabase, *filters)
-
-    def mysql_create(self, label, region, engine, ltype, **kwargs):
-        """
-        Creates an :any:`MySQLDatabase` on this account with
-        the given label, region, engine, and node type.  For example::
-
-           client = LinodeClient(TOKEN)
-
-           # look up Region and Types to use.  In this example I'm just using
-           # the first ones returned.
-           region = client.regions().first()
-           node_type = client.database.types()[0]
-           engine = client.database.engines(DatabaseEngine.engine == 'mysql')[0]
-
-           new_database = client.database.mysql_create(
-               "example-database",
-               region,
-               engine.id,
-               type.id
-            )
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-mysql-database-create
-
-        :param label: The name for this cluster
-        :type label: str
-        :param region: The region to deploy this cluster in
-        :type region: str or Region
-        :param engine: The engine to deploy this cluster with
-        :type engine: str or Engine
-        :param ltype: The Linode Type to use for this cluster
-        :type ltype: str or Type
-        """
-
-        params = {
-            "label": label,
-            "region": region.id if issubclass(type(region), Base) else region,
-            "engine": engine.id if issubclass(type(engine), Base) else engine,
-            "type": ltype.id if issubclass(type(ltype), Base) else ltype,
-        }
-        params.update(kwargs)
-
-        result = self.client.post("/databases/mysql/instances", data=params)
-
-        if "id" not in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating MySQL Database", json=result
-            )
-
-        d = MySQLDatabase(self.client, result["id"], result)
-        return d
-
-    def postgresql_instances(self, *filters):
-        """
-        Returns a list of Managed PostgreSQL Databases active on this account.
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-postgresql-databases-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of PostgreSQL databases that matched the query.
-        :rtype: PaginatedList of PostgreSQLDatabase
-        """
-        return self.client._get_and_filter(PostgreSQLDatabase, *filters)
-
-    def postgresql_create(self, label, region, engine, ltype, **kwargs):
-        """
-        Creates an :any:`PostgreSQLDatabase` on this account with
-        the given label, region, engine, and node type.  For example::
-
-           client = LinodeClient(TOKEN)
-
-           # look up Region and Types to use.  In this example I'm just using
-           # the first ones returned.
-           region = client.regions().first()
-           node_type = client.database.types()[0]
-           engine = client.database.engines(DatabaseEngine.engine == 'postgresql')[0]
-
-           new_database = client.database.postgresql_create(
-               "example-database",
-               region,
-               engine.id,
-               type.id
-            )
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-postgresql-database-create
-
-        :param label: The name for this cluster
-        :type label: str
-        :param region: The region to deploy this cluster in
-        :type region: str or Region
-        :param engine: The engine to deploy this cluster with
-        :type engine: str or Engine
-        :param ltype: The Linode Type to use for this cluster
-        :type ltype: str or Type
-        """
-
-        params = {
-            "label": label,
-            "region": region.id if issubclass(type(region), Base) else region,
-            "engine": engine.id if issubclass(type(engine), Base) else engine,
-            "type": ltype.id if issubclass(type(ltype), Base) else ltype,
-        }
-        params.update(kwargs)
-
-        result = self.client.post(
-            "/databases/postgresql/instances", data=params
-        )
-
-        if "id" not in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating PostgreSQL Database",
-                json=result,
-            )
-
-        d = PostgreSQLDatabase(self.client, result["id"], result)
-        return d
-
-    def mongodb_instances(self, *filters):
-        """
-        Returns a list of Managed MongoDB Databases active on this account.
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-mongodb-databases-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of MongoDB databases that matched the query.
-        :rtype: PaginatedList of MongoDBDatabase
-        """
-        return self.client._get_and_filter(MongoDBDatabase, *filters)
-
-    def mongodb_create(self, label, region, engine, ltype, **kwargs):
-        """
-        Creates an :any:`MongoDBDatabase` on this account with
-        the given label, region, engine, and node type.  For example::
-
-           client = LinodeClient(TOKEN)
-
-           # look up Region and Types to use.  In this example I'm just using
-           # the first ones returned.
-           region = client.regions().first()
-           node_type = client.database.types()[0]
-           engine = client.database.engines(DatabaseEngine.engine == 'mongodb')[0]
-
-           new_database = client.database.mongodb_create(
-               "example-database",
-               region,
-               engine.id,
-               type.id
-            )
-
-        API Documentation: https://www.linode.com/docs/api/databases/#managed-mongodb-database-create
-
-        :param label: The name for this cluster
-        :type label: str
-        :param region: The region to deploy this cluster in
-        :type region: str or Region
-        :param engine: The engine to deploy this cluster with
-        :type engine: str or Engine
-        :param ltype: The Linode Type to use for this cluster
-        :type ltype: str or Type
-        """
-
-        params = {
-            "label": label,
-            "region": region.id if issubclass(type(region), Base) else region,
-            "engine": engine.id if issubclass(type(engine), Base) else engine,
-            "type": ltype.id if issubclass(type(ltype), Base) else ltype,
-        }
-        params.update(kwargs)
-
-        result = self.client.post("/databases/mongodb/instances", data=params)
-
-        if "id" not in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating MongoDB Database",
-                json=result,
-            )
-
-        d = MongoDBDatabase(self.client, result["id"], result)
-        return d
 
 
 class LinodeClient:
@@ -2040,6 +105,24 @@ class LinodeClient:
 
         #: Access methods related to Managed Databases - see :any:`DatabaseGroup` for more information.
         self.database = DatabaseGroup(self)
+
+        #: Access methods related to NodeBalancers - see :any:`NodeBalancerGroup` for more information.
+        self.nodebalancers = NodeBalancerGroup(self)
+
+        #: Access methods related to Domains - see :any:`DomainGroup` for more information.
+        self.domains = DomainGroup(self)
+
+        #: Access methods related to Tags - See :any:`TagGroup` for more information.
+        self.tags = TagGroup(self)
+
+        #: Access methods related to Volumes - See :any:`VolumeGroup` for more information.
+        self.volumes = VolumeGroup(self)
+
+        #: Access methods related to Regions - See :any:`RegionGroup` for more information.
+        self.regions = RegionGroup(self)
+
+        #: Access methods related to Images - See :any:`ImageGroup` for more information.
+        self.images = ImageGroup(self)
 
     @property
     def _user_agent(self):
@@ -2205,244 +288,49 @@ class LinodeClient:
     def delete(self, *args, **kwargs):
         return self._api_call(*args, method=self.session.delete, **kwargs)
 
-    # ungrouped list functions
-    def regions(self, *filters):
-        """
-        Returns the available Regions for Linode products.
-
-        API Documentation: https://www.linode.com/docs/api/regions/#regions-list
-
-        :param filters: Any number of filters to apply to the query.
-
-        :returns: A list of available Regions.
-        :rtype: PaginatedList of Region
-        """
-        return self._get_and_filter(Region, *filters)
-
-    def images(self, *filters):
-        """
-        Retrieves a list of available Images, including public and private
-        Images available to the acting user.  You can filter this query to
-        retrieve only Images relevant to a specific query, for example::
-
-           debian_images = client.images(
-               Image.vendor == "debain")
-
-        API Documentation: https://www.linode.com/docs/api/images/#images-list
-
-        :param filters: Any number of filters to apply to the query.
-
-        :returns: A list of available Images.
-        :rtype: PaginatedList of Image
-        """
-        return self._get_and_filter(Image, *filters)
-
     def image_create(self, disk, label=None, description=None):
         """
-        Creates a new Image from a disk you own.
-
-        API Documentation: https://www.linode.com/docs/api/images/#image-create
-
-        :param disk: The Disk to imagize.
-        :type disk: Disk or int
-        :param label: The label for the resulting Image (defaults to the disk's
-                      label.
-        :type label: str
-        :param description: The description for the new Image.
-        :type description: str
-
-        :returns: The new Image.
-        :rtype: Image
+        .. note:: This method is an alias to maintain backwards compatibility.
+                  Please use :meth:`LinodeClient.images.create(...) <.ImageGroup.create>` for all new projects.
         """
-        params = {
-            "disk_id": disk.id if issubclass(type(disk), Base) else disk,
-        }
-
-        if label is not None:
-            params["label"] = label
-
-        if description is not None:
-            params["description"] = description
-
-        result = self.post("/images", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating an Image from disk {}".format(
-                    disk
-                )
-            )
-
-        return Image(self, result["id"], result)
+        return self.images.create(disk, label=label, description=description)
 
     def image_create_upload(
         self, label: str, region: str, description: str = None
     ) -> Tuple[Image, str]:
         """
-        Creates a new Image and returns the corresponding upload URL.
-
-        API Documentation: https://www.linode.com/docs/api/images/#image-upload
-
-        :param label: The label of the Image to create.
-        :type label: str
-        :param region: The region to upload to. Once the image has been created, it can be used in any region.
-        :type region: str
-        :param description: The description for the new Image.
-        :type description: str
-
-        :returns: A tuple containing the new image and the image upload URL.
-        :rtype: (Image, str)
+        .. note:: This method is an alias to maintain backwards compatibility.
+                  Please use :meth:`LinodeClient.images.create_upload(...) <.ImageGroup.create_upload>`
+                  for all new projects.
         """
-        params = {"label": label, "region": region, "description": description}
 
-        result = self.post("/images/upload", data=drop_null_keys(params))
-
-        if "image" not in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating an Image upload URL"
-            )
-
-        result_image = result["image"]
-        result_url = result["upload_to"]
-
-        return Image(self, result_image["id"], result_image), result_url
+        return self.images.create_upload(label, region, description=description)
 
     def image_upload(
         self, label: str, region: str, file: BinaryIO, description: str = None
     ) -> Image:
         """
-        Creates and uploads a new image.
-
-        API Documentation: https://www.linode.com/docs/api/images/#image-upload
-
-        :param label: The label of the Image to create.
-        :type label: str
-        :param region: The region to upload to. Once the image has been created, it can be used in any region.
-        :type region: str
-        :param file: The BinaryIO object to upload to the image. This is generally obtained from open("myfile", "rb").
-        :param description: The description for the new Image.
-        :type description: str
-
-        :returns: The resulting image.
-        :rtype: Image
+        .. note:: This method is an alias to maintain backwards compatibility.
+                  Please use :meth:`LinodeClient.images.upload(...) <.ImageGroup.upload>` for all new projects.
         """
-
-        image, url = self.image_create_upload(
-            label, region, description=description
-        )
-
-        requests.put(
-            url,
-            headers={"Content-Type": "application/octet-stream"},
-            data=file,
-        )
-
-        image._api_get()
-
-        return image
-
-    def domains(self, *filters):
-        """
-        Retrieves all of the Domains the acting user has access to.
-
-        API Documentation: https://www.linode.com/docs/api/domains/#domains-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Domains the acting user can access.
-        :rtype: PaginatedList of Domain
-        """
-        return self._get_and_filter(Domain, *filters)
-
-    def nodebalancers(self, *filters):
-        """
-        Retrieves all of the NodeBalancers the acting user has access to.
-
-        API Documentation: https://www.linode.com/docs/api/nodebalancers/#nodebalancers-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of NodeBalancers the acting user can access.
-        :rtype: PaginatedList of NodeBalancers
-        """
-        return self._get_and_filter(NodeBalancer, *filters)
+        return self.images.upload(label, region, file, description=description)
 
     def nodebalancer_create(self, region, **kwargs):
         """
-        Creates a new NodeBalancer in the given Region.
-
-        API Documentation: https://www.linode.com/docs/api/nodebalancers/#nodebalancer-create
-
-        :param region: The Region in which to create the NodeBalancer.
-        :type region: Region or str
-
-        :returns: The new NodeBalancer
-        :rtype: NodeBalancer
+        .. note:: This method is an alias to maintain backwards compatibility.
+                  Please use
+                  :meth:`LinodeClient.nodebalancers.create(...) <.NodeBalancerGroup.create>`
+                  for all new projects.
         """
-        params = {
-            "region": region.id if isinstance(region, Base) else region,
-        }
-        params.update(kwargs)
-
-        result = self.post("/nodebalancers", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Nodebalaner!", json=result
-            )
-
-        n = NodeBalancer(self, result["id"], result)
-        return n
+        return self.nodebalancers.create(region, **kwargs)
 
     def domain_create(self, domain, master=True, **kwargs):
         """
-        Registers a new Domain on the acting user's account.  Make sure to point
-        your registrar to Linode's nameservers so that Linode's DNS manager will
-        correctly serve your domain.
-
-        API Documentation: https://www.linode.com/docs/api/domains/#domain-create
-
-        :param domain: The domain to register to Linode's DNS manager.
-        :type domain: str
-        :param master: Whether this is a master (defaults to true)
-        :type master: bool
-        :param tags: A list of tags to apply to the new domain.  If any of the
-                     tags included do not exist, they will be created as part of
-                     this operation.
-        :type tags: list[str]
-
-        :returns: The new Domain object.
-        :rtype: Domain
+        .. note:: This method is an alias to maintain backwards compatibility.
+                  Please use :meth:`LinodeClient.domains.create(...) <.DomainGroup.create>` for all
+                  new projects.
         """
-        params = {
-            "domain": domain,
-            "type": "master" if master else "slave",
-        }
-        params.update(kwargs)
-
-        result = self.post("/domains", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Domain!", json=result
-            )
-
-        d = Domain(self, result["id"], result)
-        return d
-
-    def tags(self, *filters):
-        """
-        Retrieves the Tags on your account.  This may only be attempted by
-        unrestricted users.
-
-        API Documentation: https://www.linode.com/docs/api/domains/#domain-create
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Tags on the account.
-        :rtype: PaginatedList of Tag
-        """
-        return self._get_and_filter(Tag, *filters)
+        return self.domains.create(domain, master=master, **kwargs)
 
     def tag_create(
         self,
@@ -2454,146 +342,26 @@ class LinodeClient:
         entities=[],
     ):
         """
-        Creates a new Tag and optionally applies it to the given entities.
-
-        API Documentation: https://www.linode.com/docs/api/tags/#tags-list
-
-        :param label: The label for the new Tag
-        :type label: str
-        :param entities: A list of objects to apply this Tag to upon creation.
-                         May only be taggable types (Linode Instances, Domains,
-                         NodeBalancers, or Volumes).  These are applied *in addition
-                         to* any IDs specified with ``instances``, ``domains``,
-                         ``nodebalancers``, or ``volumes``, and is a convenience
-                         for sending multiple entity types without sorting them
-                         yourself.
-        :type entities: list of Instance, Domain, NodeBalancer, and/or Volume
-        :param instances: A list of Linode Instances to apply this Tag to upon
-                        creation
-        :type instances: list of Instance or list of int
-        :param domains: A list of Domains to apply this Tag to upon
-                        creation
-        :type domains: list of Domain or list of int
-        :param nodebalancers: A list of NodeBalancers to apply this Tag to upon
-                        creation
-        :type nodebalancers: list of NodeBalancer or list of int
-        :param volumes: A list of Volumes to apply this Tag to upon
-                        creation
-        :type volumes: list of Volumes or list of int
-
-        :returns: The new Tag
-        :rtype: Tag
+        .. note:: This method is an alias to maintain backwards compatibility.
+                  Please use :meth:`LinodeClient.tags.create(...) <.TagGroup.create>` for all new projects.
         """
-        linode_ids, nodebalancer_ids, domain_ids, volume_ids = [], [], [], []
-
-        # filter input into lists of ids
-        sorter = zip(
-            (linode_ids, nodebalancer_ids, domain_ids, volume_ids),
-            (instances, nodebalancers, domains, volumes),
+        return self.tags.create(
+            label,
+            instances=instances,
+            domains=domains,
+            nodebalancers=nodebalancers,
+            volumes=volumes,
+            entities=entities,
         )
-
-        for id_list, input_list in sorter:
-            # if we got something, we need to find its ID
-            if input_list is not None:
-                for cur in input_list:
-                    if isinstance(cur, int):
-                        id_list.append(cur)
-                    else:
-                        id_list.append(cur.id)
-
-        # filter entities into id lists too
-        type_map = {
-            Instance: linode_ids,
-            NodeBalancer: nodebalancer_ids,
-            Domain: domain_ids,
-            Volume: volume_ids,
-        }
-
-        for e in entities:
-            if type(e) in type_map:
-                type_map[type(e)].append(e.id)
-            else:
-                raise ValueError("Unsupported entity type {}".format(type(e)))
-
-        # finally, omit all id lists that are empty
-        params = {
-            "label": label,
-            "linodes": linode_ids or None,
-            "nodebalancers": nodebalancer_ids or None,
-            "domains": domain_ids or None,
-            "volumes": volume_ids or None,
-        }
-
-        result = self.post("/tags", data=params)
-
-        if not "label" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating Tag!", json=result
-            )
-
-        t = Tag(self, result["label"], result)
-        return t
-
-    def volumes(self, *filters):
-        """
-        Retrieves the Block Storage Volumes your user has access to.
-
-        API Documentation: https://www.linode.com/docs/api/volumes/#volumes-list
-
-        :param filters: Any number of filters to apply to this query.
-
-        :returns: A list of Volumes the acting user can access.
-        :rtype: PaginatedList of Volume
-        """
-        return self._get_and_filter(Volume, *filters)
 
     def volume_create(self, label, region=None, linode=None, size=20, **kwargs):
         """
-        Creates a new Block Storage Volume, either in the given Region or
-        attached to the given Instance.
-
-        API Documentation: https://www.linode.com/docs/api/volumes/#volumes-list
-
-        :param label: The label for the new Volume.
-        :type label: str
-        :param region: The Region to create this Volume in.  Not required if
-                       `linode` is provided.
-        :type region: Region or str
-        :param linode: The Instance to attach this Volume to.  If not given, the
-                       new Volume will not be attached to anything.
-        :type linode: Instance or int
-        :param size: The size, in GB, of the new Volume.  Defaults to 20.
-        :type size: int
-        :param tags: A list of tags to apply to the new volume.  If any of the
-                     tags included do not exist, they will be created as part of
-                     this operation.
-        :type tags: list[str]
-
-        :returns: The new Volume.
-        :rtype: Volume
+        .. note:: This method is an alias to maintain backwards compatibility.
+                  Please use :meth:`LinodeClient.volumes.create(...) <.VolumeGroup.create>` for all new projects.
         """
-        if not (region or linode):
-            raise ValueError("region or linode required!")
-
-        params = {
-            "label": label,
-            "size": size,
-            "region": region.id if issubclass(type(region), Base) else region,
-            "linode_id": linode.id
-            if issubclass(type(linode), Base)
-            else linode,
-        }
-        params.update(kwargs)
-
-        result = self.post("/volumes", data=params)
-
-        if not "id" in result:
-            raise UnexpectedResponseError(
-                "Unexpected response when creating volume!", json=result
-            )
-
-        v = Volume(self, result["id"], result)
-        return v
+        return self.volumes.create(
+            label, region=region, linode=linode, size=size, **kwargs
+        )
 
     # helper functions
     def _get_and_filter(self, obj_type, *filters):
